Tensorflow 2.0 Base Added
Added in a basis for the workbench in Tensorflow 2.0 going foreward. Utilizes a YOLO graph and features a script for converting the folder that the system saves the model to into coreml. Has similar functionality to the previous 1.0 workbench for display functionality, image detection, and so on. Capable of processing large datasets such as those found on the Voco database site and thus should have no problem handling data for dice so long as it is formatted similarly to the datasets found on the Voco site.