Added .h5 save and fixed tflite conversion
Note: Tflie does not support custom functions such as combne non max suppresion which is used in the yolo_nms() function in model.py. We must either figure out how to  get a working "FlexDelegate" or to remove the use of nms. Also the tflite models produced are still very large and thus we made need to get yolo_lite to work to make the models less large, or mess with the tf lite converter more