Stable version for workbench
coreml and tflite converson should both be fully functional. Tflte produces models with an output of [1, 1000] which has yet to be figured out how to have t used in android apps, if converter.allow_custom_ops is set to True then the output shape wll be more favorable, but the system will have also included the custom function for nms which for the interpreter need a delegate and have very sparse support. Alos added more verbose error messaging to conversion portion of the code